data fetcher/requirements.txt
requests
pandas
scikit-learn
joblib
mysql-connector-python
flask
